+++ .automation/test/golang/golang_bad_01.go
@@ -1,3 +1,7 @@
+package main
+
 if len(in) == 0 {
-  return "", fmt.Errorf("Input is empty")
+	return "", fmt.Errorf("Input is empty")
 }
+
+sum(1, 2)