+++ .automation/test/golang/golang_good_01.go
@@ -4,4 +4,5 @@ import "fmt"
 
 func main() {
 	fmt.Println("hello world")
+	sum(1, 2)
 }