+++ .automation/test/golang/golang_good_maths_01.go
@@ -1,0 +1,5 @@
+package main
+
+func sum(a int, b int) int {
+	return a + b
+}